Remove default epsg from settings. Include when create schema in data management
