+++ asistente_ladm_col/ui/qgis_model_baker/dlg_import_schema.ui
@@ -144,7 +144,7 @@
           </sizepolicy>
          </property>
          <property name="title">
-          <string>PostgreSQL</string>
+          <string>Database settings</string>
          </property>
          <layout class="QGridLayout" name="gridLayout">
           <item row="0" column="1">
@@ -178,6 +178,35 @@
             </property>
            </widget>
           </item>
+          <item row="1" column="1">
+           <widget class="QgsProjectionSelectionWidget" name="crsSelector" native="true">
+            <property name="sizePolicy">
+             <sizepolicy hsizetype="MinimumExpanding" vsizetype="Preferred">
+              <horstretch>0</horstretch>
+              <verstretch>0</verstretch>
+             </sizepolicy>
+            </property>
+            <property name="focusPolicy">
+             <enum>Qt::StrongFocus</enum>
+            </property>
+           </widget>
+          </item>
+          <item row="1" column="0">
+           <widget class="QLabel" name="crs_label">
+            <property name="font">
+             <font>
+              <weight>50</weight>
+              <bold>false</bold>
+             </font>
+            </property>
+            <property name="toolTip">
+             <string>Coordinate Reference System</string>
+            </property>
+            <property name="text">
+             <string>Default CRS:</string>
+            </property>
+           </widget>
+          </item>
          </layout>
         </widget>
        </item>
@@ -192,6 +221,12 @@
   <customwidget>
    <class>QgsCollapsibleGroupBox</class>
    <extends>QGroupBox</extends>
+   <header>qgis.gui</header>
+   <container>1</container>
+  </customwidget>
+  <customwidget>
+   <class>QgsProjectionSelectionWidget</class>
+   <extends>QWidget</extends>
    <header>qgis.gui</header>
    <container>1</container>
   </customwidget>
